+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Unreleased
 - Converted repository to a Cargo workspace
-- Implemented `SmolBytes` backed by inline storage (up to 39 bytes) or `bytes::Bytes`
+- Implemented `Bytes` backed by inline storage (up to 39 bytes) or `bytes::Bytes`